requirements/prod.txt
# Django stuff
-e git://github.com/django/django@36c82ac8#egg=django

# Templates
-e git://github.com/jbalogh/jingo.git#egg=jingo
-e git://github.com/jsocol/jingo-minify.git#egg=jingo-minify
GitPython==0.1.7

# Various tidbits
-e git://github.com/jsocol/commonware.git#egg=commonware
-e git://github.com/mozilla/nuggets.git#egg=nuggets
-e git://github.com/django-extensions/django-extensions.git#egg=django-extensions

# Security
-e git://github.com/fwenzel/django-sha2.git#egg=django-sha2
-e git://github.com/jsocol/bleach.git#egg=bleach

# Celery: Message queue
celery
django-celery

# L10n
Babel>=0.9.4
-e git://github.com/clouserw/tower.git#egg=tower
-e git://github.com/fwenzel/django-mozilla-product-details#egg=django-mozilla-product-details

# Auth
-e hg+https://code.google.com/p/httplib2/@6525cadfde53#egg=httplib2
-e git://github.com/mozilla/django-browserid.git#egg=django-browserid

# Apps
-e git://github.com/lmorchard/django-badger.git#egg=django-badger
-e git://github.com/lmorchard/django-badger-multiplayer.git#egg=django-badger-multiplayer
-e git://github.com/lmorchard/django-badger-api.git#egg=django-badger-api
